+++ doc.go
@@ -1,4 +1,2 @@
-/*
-汉语拼音转换工具。
-*/
+// Package pinyin : 汉语拼音转换工具
 package pinyin