+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.2.0 (yyyy-mm-dd)
-** 不兼容之前的版本 **
+**不兼容之前的版本**
 
 * 新增 ``func NewArgs() *Args``
 * 各函数内的 ``a Args`` 参数改为 ``a *Args``