+++ CHANGELOG.md
@@ -1,10 +1,8 @@
 # Changelog
 
 ## 0.2.0 (yyyy-mm-dd)
-**不兼容之前的版本**
 
-* 新增 ``func NewArgs() *Args``
-* 各函数内的 ``a Args`` 参数改为 ``a *Args``
+* 新增 ``func NewArgs() Args``
 * 解决 ``Args.Separator`` 无法赋值为 ``""`` 的 BUG
 * 规范命名:
     * ``NORMAL`` -> ``Normal``